Move error types and functions Cardano.CLI.Legacy.Run.StakeAddress to Cardano.CLI.EraBased.Errors.StakeAddress
