+++ cardano-cli/cardano-cli.cabal
@@ -71,6 +71,7 @@ library
                         Cardano.CLI.EraBased.Commands.Governance.Actions
                         Cardano.CLI.EraBased.Commands.Governance.Committee
                         Cardano.CLI.EraBased.Errors.Common
+                        Cardano.CLI.EraBased.Errors.StakeAddress
                         Cardano.CLI.EraBased.Options.Common
                         Cardano.CLI.EraBased.Options.Governance
                         Cardano.CLI.EraBased.Options.Governance.Actions
